requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tqdm>=4.65.0
faiss-cpu>=1.7.4  # or faiss-gpu for GPU support

# Deep Learning & Transformers
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.2
accelerate>=0.20.0

# LangChain and LLM
langchain>=0.0.300
langchain-openai>=0.0.2
langchain-community>=0.0.10
langchain-core>=0.1.0

# FastAPI and Web
fastapi>=0.100.0
uvicorn>=0.23.0
requests>=2.31.0
pydantic>=2.0.0

# Data Processing
datasets>=2.14.0
rank-bm25>=0.2.2
llama-index>=0.8.0

# Utilities
python-dotenv>=1.0.0
typing-extensions>=4.7.0
importlib-metadata>=4.6.4
